[LargestContentfulPaint] Replace responseEnd with loadTime
This CL replaces responseEnd with loadTime. Before, responseEnd would be
used and would be problematic for memory cached images. We change it to
loadTime, the time at which LayoutObject::ImageNotifyFinished is called.
This time has the advantage of being independent of the memory cache and
being after the image has been appended to the DOM, as shown by a test.

To implement loadTime, we use a new HashMap. This HashMap populates the
timestamps that will be used in ImageRecords. We cannot use the existing
data structures because this happens before we know whether the image
will be added to the set of visible or invisible images, etc.

Bug: 982046, 965505
Change-Id: I50d4f304a35b60409c58c54a5bd912e7d15825e9