Fixing rendering of backdrop-filter with opacity
Prior to this CL, a node that contained both opacity and backdrop-filter
would sometimes not be properly rendered. The opacity was, in some cases,
applied as if there was an opaque black background behind the filter
region, rather than compositing the filtered image into the background
with the specified opacity. This CL changes the way opacity is applied,
moving it to the paint of the final filtered image, which fixes the
problem.

Bug: 983252
Change-Id: I9fae88c8f7b1c950d5616e1c9eeaaa664bab649a